+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 + [x] Hamburger: updates for using props, add docs
 + [x] ButtonGroup: updates for using props, add docs
 + [x] Select: add focus to input when dropdown
++ [x] Textarea: fix autosize jumping, issue 1400 from core Metro 4
 
 ### 0.1.0
 + [x] Initial version, contains 40 components, common styles, additional colors, routines.